+++ Src/WebAPI/Controllers/ColorController.cs
@@ -5,6 +5,7 @@ using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using WebAPI.DTOs.ColorDtos;
 
@@ -28,28 +29,39 @@ namespace WebAPI.Controllers
             return Ok(await _unitOfWork.ColorService.GetAllColor());
         }
 
-        [Authorize(Roles = "admin")]
+        //[Authorize(Roles = "admin")]
         [HttpPost]
+        [Authorize]
         public async Task<IActionResult> CreateColor([FromBody] CreateColorDto ColorDto)
         {
-
-            if (!ModelState.IsValid)
+            string Roletype = User.FindFirstValue("Roletype");
+            if (Roletype == "False")
             {
-                return BadRequest("Invalid data.");
+                if (!ModelState.IsValid)
+                {
+                    return BadRequest("Invalid data.");
+                }
+                await _unitOfWork.ColorService.AddColor(ColorDto.ColorName);
+                _unitOfWork.Complete();
+                return Ok();
             }
-
-            await _unitOfWork.ColorService.AddColor(ColorDto.ColorName);
-            _unitOfWork.Complete();
-            return Ok();
+            return Unauthorized();
+            
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
-        [Authorize(Roles = "admin")]
         public async Task<IActionResult> DeleteColor(int id)
         {
-            await _unitOfWork.ColorService.DeleteColor(id);
-            _unitOfWork.Complete();
-            return NoContent();
+            string Roletype = User.FindFirstValue("Roletype");
+            if (Roletype == "False")
+            {
+                await _unitOfWork.ColorService.DeleteColor(id);
+                _unitOfWork.Complete();
+                return NoContent();
+            }
+            return Unauthorized();
+
         }
 
     }
